redis locking and refactor complete
