+++ internal/server/move-cheats.go
@@ -9,19 +9,11 @@ import (
 	"github.com/JonathanWinters/go_test/internal/core"
 )
 
-type MoveRequestBody struct {
-	PrimaryKey int
-	Move       int
-	GodMode    bool
-}
-
 var MoveRequestQueue = []core.MoveRequest{}
 
-// !FIXED
-// !INFO EFC: for build flag code, try to isolate only the meaningful differences, reduce dupe code
 func HandleMove(w http.ResponseWriter, r *http.Request) {
 
-	mrb, err := DecodeJson(w, r)
+	mrb, err := DecodeMoveJson(w, r)
 	if err != nil {
 		return
 	}
@@ -31,8 +23,6 @@ func HandleMove(w http.ResponseWriter, r *http.Request) {
 		Move:       mrb.Move,
 		GodMode:    mrb.GodMode,
 	}
-
-	log.Println("Decode Successful")
 
 	rawResult, err := ProcessMoveRequest(w, moveRequest)
 	if err != nil {
