+++ internal/server/submit-macos-linux.go
@@ -3,16 +3,8 @@
 package server
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
 	"log"
 	"net/http"
-	"strings"
-
-	"github.com/JonathanWinters/go_test/internal/core"
-	"github.com/JonathanWinters/go_test/internal/definitions"
 )
 
 type RequestBody struct {
@@ -22,133 +14,15 @@ type RequestBody struct {
 
 func HandleSubmit(w http.ResponseWriter, r *http.Request) {
 
-	// If the Content-Type header is present, check that it has the value
-	// application/json. Note that we parse and normalize the header to remove
-	// any additional parameters (like charset or boundary information) and normalize
-	// it by stripping whitespace and converting to lowercase before we check the
-	// value.
-	ct := r.Header.Get("Content-Type")
-	if ct != "" {
-		mediaType := strings.ToLower(strings.TrimSpace(strings.Split(ct, ";")[0]))
-		if mediaType != "application/json" {
-			msg := "Content-Type header is not application/json"
-			log.Println(msg)
-			http.Error(w, msg, http.StatusUnsupportedMediaType)
-			return
-		}
-	}
-
-	// Use http.MaxBytesReader to enforce a maximum read of 1MB from the
-	// response body. A request body larger than that will now result in
-	// Decode() returning a "http: request body too large" error.
-	r.Body = http.MaxBytesReader(w, r.Body, 1048576)
-
-	// Setup the decoder and call the DisallowUnknownFields() method on it.
-	// This will cause Decode() to return a "json: unknown field ..." error
-	// if it encounters any extra unexpected fields in the JSON. Strictly
-	// speaking, it returns an error for "keys which do not match any
-	// non-ignored, exported fields in the destination".
-	dec := json.NewDecoder(r.Body)
-	dec.DisallowUnknownFields()
-
-	var p RequestBody
-	err := dec.Decode(&p)
+	rawResult, err := ProcessSubmitRequest(w, r)
 	if err != nil {
-		var syntaxError *json.SyntaxError
-		var unmarshalTypeError *json.UnmarshalTypeError
-		var maxBytesError *http.MaxBytesError
-
-		switch {
-		// Catch any syntax errors in the JSON and send an error message
-		// which interpolates the location of the problem to make it
-		// easier for the client to fix.
-		case errors.As(err, &syntaxError):
-			msg := fmt.Sprintf("Request body contains badly-formed JSON (at position %d)", syntaxError.Offset)
-			log.Println(msg)
-			http.Error(w, msg, http.StatusBadRequest)
-
-		// In some circumstances Decode() may also return an
-		// io.ErrUnexpectedEOF error for syntax errors in the JSON. There
-		// is an open issue regarding this at
-		// https://github.com/golang/go/issues/25956.
-		case errors.Is(err, io.ErrUnexpectedEOF):
-			msg := "Request body contains badly-formed JSON"
-			log.Println(msg)
-			http.Error(w, msg, http.StatusBadRequest)
-
-		// Catch any type errors, like trying to assign a string in the
-		// JSON request body to a int field in our Person struct. We can
-		// interpolate the relevant field name and position into the error
-		// message to make it easier for the client to fix.
-		case errors.As(err, &unmarshalTypeError):
-			msg := fmt.Sprintf("Request body contains an invalid value for the %q field (at position %d)", unmarshalTypeError.Field, unmarshalTypeError.Offset)
-			log.Println(msg)
-			http.Error(w, msg, http.StatusBadRequest)
-
-		// Catch the error caused by extra unexpected fields in the request
-		// body. We extract the field name from the error message and
-		// interpolate it in our custom error message. There is an open
-		// issue at https://github.com/golang/go/issues/29035 regarding
-		// turning this into a sentinel error.
-		case strings.HasPrefix(err.Error(), "json: unknown field "):
-			fieldName := strings.TrimPrefix(err.Error(), "json: unknown field ")
-			msg := fmt.Sprintf("Request body contains unknown field %s", fieldName)
-			log.Println(msg)
-			http.Error(w, msg, http.StatusBadRequest)
-
-		// An io.EOF error is returned by Decode() if the request body is
-		// empty.
-		case errors.Is(err, io.EOF):
-			msg := "Request body must not be empty"
-			log.Println(msg)
-			http.Error(w, msg, http.StatusBadRequest)
-
-		// Catch the error caused by the request body being too large.
-		case errors.As(err, &maxBytesError):
-			msg := fmt.Sprintf("Request body must not be larger than %d bytes", maxBytesError.Limit)
-			log.Println(msg)
-			http.Error(w, msg, http.StatusRequestEntityTooLarge)
-
-		// Otherwise default to logging the error and sending a 500 Internal
-		// Server Error response.
-		default:
-			log.Println(err.Error())
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		}
+		log.Fatal(err)
 		return
 	}
 
-	// Call decode again, using a pointer to an empty anonymous struct as
-	// the destination. If the request body only contained a single JSON
-	// object this will return an io.EOF error. So if we get anything else,
-	// we know that there is additional data in the request body.
-	// err = dec.Decode(&struct{}{})
-	// if !errors.Is(err, io.EOF) {
-	// 	msg := "Request body must only contain a single JSON object"
-	// 	http.Error(w, msg, http.StatusBadRequest)
-	// 	return
-	// }
-
-	log.Println("Submit Decode Successful")
-
-	responseObj := RequestBody{
-		UserId: p.UserId,
-		Level:  p.Level,
-	}
-
-	userid := definitions.UserIDFromString(responseObj.UserId)
-
-	submitRequest := core.SubmitRequest{
-		UserID: userid,
-		Level:  responseObj.Level,
-	}
-
-	submitResponse := core.HandleSubmit(w, submitRequest)
-
-	rawResult, err := json.Marshal(submitResponse)
+	_, err = w.Write(rawResult)
 	if err != nil {
+		log.Fatal(err)
 		return
 	}
-
-	fmt.Fprintf(w, "%s", rawResult)
 }
