--- scripts/endpoints/ids.json
@@ -1,8 +1,0 @@
-{
-    "1": {
-        "id": [
-            "29a61936-7811-4db4-8d99-5662370e043d",
-            ""
-        ]
-    }
-}
