--- tmp/move.json
@@ -1,21 +1,0 @@
-{
-  "Error": "",
-  "Result": "Move Not Allowed: Wall Hit",
-  "PlayerHitPoints": 1,
-  "Position": {
-    "X": 1,
-    "Y": 1
-  },
-  "LatestMap": [
-    [1,1,1,1,0,1,1,1],
-    [1,4,0,0,0,0,0,1],
-    [1,0,1,1,1,3,1,1],
-    [1,0,0,0,1,0,2,1],
-    [1,1,1,0,1,1,0,1],
-    [1,0,0,0,1,0,0,1],
-    [1,0,1,1,1,0,1,1],
-    [1,0,0,0,0,0,0,1],
-    [1,1,1,1,1,1,1,1]
-  ]
-}
-
