--- tmp/submit.json
@@ -1,21 +1,0 @@
-{
-  "Error": "",
-  "PrimaryKey": 1,
-  "LevelID": "f1ca4ac9-0e75-469e-a108-cab2932549bb",
-  "Map": [
-    [1,1,1,1,0,1,1,1],
-    [1,0,0,0,0,0,0,1],
-    [1,0,1,1,1,3,1,1],
-    [1,0,0,0,1,0,2,1],
-    [1,1,1,0,1,1,0,1],
-    [1,0,0,0,1,0,0,1],
-    [1,0,1,1,1,0,1,1],
-    [1,0,0,4,0,0,0,1],
-    [1,1,1,1,1,1,1,1]
-  ],
-  "Position": {
-    "X": 3,
-    "Y": 7
-  }
-}
-
